[back:winnative] Fix lost of iface infos when an other is added/removed
When an interface is added or removed, we previously clear the map and rebuild it, then a comparaison was done with previous map to properly emit signals.

But doing this implies that previous retrieved interfaces lost some of their informations (like list of networks).
So keep reference of interface already known if not changed